move entities from ./schema to ./entities
