+++ app/entities/User.py
@@ -1,16 +1,10 @@
-import enum
 from typing import Self
 
 from pydantic import BaseModel, EmailStr, model_validator
 from ulid import ULID
 
 from app.auth.hash import generate_password_hash
-
-
-class Role(enum.Enum):
-    ADMIN = 'admin'
-    USER = 'user'
-    GUEST = 'guest'
+from app.database.models import Role, UserDB
 
 
 class UserCreateDTO(BaseModel):
@@ -23,7 +17,7 @@ class UserCreateDTO(BaseModel):
     def check_passwords_match(self) -> Self:
         pw1 = self.password
         pw2 = self.password2
-        if (pw1 is not None) and (pw2 is not None) and (pw1 != pw2):
+        if pw1 != pw2:
             raise ValueError('passwords does not match')
 
         self.password = generate_password_hash(pw1)
@@ -42,7 +36,7 @@ class User(BaseModel):
         return cls(email=dto.email, password=dto.password, role=dto.role)
 
     @classmethod
-    def from_db(cls, db_user):
+    def from_db(cls, db_user: UserDB):
         return cls(
             ulid=ULID.from_str(db_user.ulid),
             email=db_user.email,
