--- app/schemas/Info.py
@@ -1,6 +1,0 @@
-from pydantic import BaseModel
-
-
-class Info(BaseModel):
-    count: int = 0
-    total_pages: int = 1
