--- app/schemas/Message.py
@@ -1,5 +1,0 @@
-from pydantic import BaseModel
-
-
-class Message(BaseModel):
-    content: str
