--- app/schemas/RepositorySettings.py
@@ -1,5 +1,0 @@
-from pydantic import BaseModel
-
-
-class RepositorySettings(BaseModel):
-    pagination: int = 10
